+++ example/src/controllers/Users.ts
@@ -1,8 +1,13 @@
-import { Controller } from 'spooked';
+import { Controller, Route, Request, Response } from 'spooked';
 
 @Controller()
 class Users {
-
+    @Route('GET')
+    all(request: Request, response: Response): void {
+        response.jsonp({
+            message: 'hello',
+        });
+    }
 }
 
 export default Users;