+++ src/types/IController.ts
@@ -1,4 +1,0 @@
-export interface IController {
-    target: Function;
-    name: string;
-}