+++ src/types/server/IController.ts
@@ -1,0 +1,6 @@
+interface IController {
+    target: Function;
+    name: string;
+}
+
+export default IController;